+++ include/libpldm/transport/mctp-demux.h
@@ -21,7 +21,7 @@ void pldm_transport_mctp_demux_destroy(struct pldm_transport_mctp_demux *ctx);
 struct pldm_transport *
 pldm_transport_mctp_demux_core(struct pldm_transport_mctp_demux *ctx);
 
-#ifdef PLDM_HAS_POLL
+#if HAVE_POLL_H
 struct pollfd;
 /* Init pollfd for async calls */
 int pldm_transport_mctp_demux_init_pollfd(struct pldm_transport *t,